Take compiler flags from M.extern.<target>.cfg Ignore-this: 723580046b003e1f3450ef477876b1b3 This is a quick fix to be able to compile modules and link programs that refer to external libraries, without having to modify the compiler configuration and reinstalling the compiler.
The configuration file has the same syntax as the rts<target/timberc.cfg file.
The cCompiler field is ignored. The flag fields are appended to the flags
specified in timberc.cfg. When linking, flags from all relevant configuration
files are appended. Duplicated flags are supressed. Flags are assumed to be
words delimited by wite space.