Tolerate name clashes when substituting in expressions. Ignore-this: 987e99b4ada3d2a2510cbc8337e1d3ab Names have generally been assumed to be unique, which means that an instance Subst s Name Exp, can safely apply a substitution under lambdas and other constructions that bind variables, without risking name capture.
This patch changes the substitution functions to perform alpha conversion to
avoid name capture, if necessary. This requires checking for name clashes and
generation of fresh names. Without changing the types of the substitution
functions, the price for this is repeated traversal of substitutions and
expressions, to compute free variables. The worst case complexity, when every
binding needs alpha conversion, is exponential in the nesting depth of binders
(i.e. a subexpression under n binders will be traversed 2^n times).

Comments in the source code indicate an easy way to revert back to the old
behavior, in case this implementation causes problems.